requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
aiofiles>=23.0.0
lxml>=4.9.0
urllib3>=2.0.0
tqdm>=4.65.0
pandas>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
openai>=1.0.0
pillow>=10.0.0 